requirements.txt
Flask==0.10.1
Flask-Classy==0.6.10
Flask-Cors==2.0.1
Flask-Redis==0.1.0
Flask-Script==2.0.5
Jinja2==2.7.3
Markdown==2.6.2
MarkupSafe==0.23
PyYAML==3.11
Werkzeug==0.10.4
argparse==1.3.0
arrow==0.5.4
babelfish==0.5.4
bleach==1.4.1
fuzzywuzzy==0.5.0
guessit==0.10.3
html5lib==0.99999
itsdangerous==0.24
pbr==1.2.0
python-dateutil==2.4.2
redis==2.10.3
requests==2.7.0
six==1.9.0
stevedore==1.5.0
